requirements.txt
# TapCraft Dependencies
pyyaml>=6.0
playsound>=1.3.0

# macOS
pyobjc-framework-Quartz>=9.0; sys_platform == "darwin"
pyobjc-framework-Cocoa>=9.0; sys_platform == "darwin"
rumps>=0.4.0; sys_platform == "darwin"

# Windows
pynput>=1.7.6; sys_platform == "win32"

# Linux
evdev>=1.6.0; sys_platform == "linux"
pynput>=1.7.6; sys_platform == "linux"

# Microphone-based slap detection (cross-platform)
sounddevice>=0.4.6
numpy>=1.24.0

# Optional: for GUI configurator
# tkinter is built-in
